Move multibuffer tests to their own source file
